requirements.txt
torch
torchvision
Pillow
gdown
opencv-python
numpy
kaggle
transfomers
datasets
pandas
matplotlib
scikit-learn